# [FIX] PromQL: support variable scalar parameter in aggregations in range queries