--- a/promql/promqltest/testdata/aggregators.test
+++ b/promql/promqltest/testdata/aggregators.test
@@ -482,7 +482,8 @@ eval instant at 1m quantile without(point)((scalar(foo)), data)
 	{test="three samples"} 1.6
 	{test="uneven samples"} 2.8
 
-eval_warn instant at 1m quantile without(point)(NaN, data)
+eval instant at 1m quantile without(point)(NaN, data)
+	expect warn msg: PromQL warning: quantile value should be between 0 and 1, got NaN
     {test="two samples"} NaN
     {test="three samples"} NaN
     {test="uneven samples"} NaN