Fixes #15971