@beorn7 2025-04-09
Thanks for tackling this.

While this isn't directly un-doing https://github.com/prometheus/prometheus/pull/13744, it neither does what we sketched out in the dev-summit (find out if the parameter is constant at parse-time and take different code paths then).

But maybe the approach here is still fast enough? Have you tried running the micro benchmarks? You could compare the effect of this PR with the findings in https://github.com/prometheus/prometheus/pull/13744 .

Maybe we could also do something more in the `allParamMatchCondition` style, e.g. first finding out if the parameter doesn't change.

I would also welcome @bboreham and @machine424 to take a look here. You are way better domain experts here than me.