@beorn7 2025-04-10
Remember to run the `RangeQuery` benchmarks with steps=100 and steps=1000 as it was done in https://github.com/prometheus/prometheus/pull/13744 back then to show the improvements.