@NeerajGartia21 2025-04-10
These are the benchmarks run against this PR. First 1000 steps and then 100 steps.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql
cpu: Apple M1
                                                                                                         │ /Users/neerajgartia/Desktop/dev/cp/before_1000.txt │ /Users/neerajgartia/Desktop/dev/cp/after_1000.txt │
                                                                                                         │                       sec/op                       │           sec/op            vs base               │
RangeQuery/expr=a_one,steps=1000-8                                                                                                               35.22µ ±  8%                 35.23µ ±  1%        ~ (p=0.937 n=6)
RangeQuery/expr=a_ten,steps=1000-8                                                                                                               302.0µ ±  1%                 300.9µ ±  0%        ~ (p=0.065 n=6)
RangeQuery/expr=a_hundred,steps=1000-8                                                                                                           2.981m ± 64%                 2.996m ±  1%        ~ (p=0.310 n=6)
RangeQuery/expr=rate(a_one[1m]),steps=1000-8                                                                                                     82.58µ ±  0%                 82.62µ ±  1%        ~ (p=0.394 n=6)
RangeQuery/expr=rate(a_ten[1m]),steps=1000-8                                                                                                     776.1µ ± 12%                 775.1µ ±  0%        ~ (p=0.589 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-8                                                                                                 7.734m ±  1%                 7.774m ±  0%        ~ (p=0.180 n=6)
RangeQuery/expr=rate(a_one[1m]),steps=10000-8                                                                                                    797.6µ ±  0%                 801.7µ ±  1%   +0.52% (p=0.009 n=6)
RangeQuery/expr=rate(a_ten[1m]),steps=10000-8                                                                                                    8.159m ±  5%                 8.140m ±  0%        ~ (p=0.589 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-8                                                                                                81.56m ±  0%                 81.80m ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=rate(sparse[1m]),steps=10000-8                                                                                                   5.318m ±  5%                 5.367m ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=double_exponential_smoothing(a_one[1d],_0.3,_0.3),steps=1000-8                                                                   58.36m ±  0%                 58.36m ±  0%        ~ (p=0.699 n=6)
RangeQuery/expr=double_exponential_smoothing(a_ten[1d],_0.3,_0.3),steps=1000-8                                                                   582.2m ±  0%                 583.1m ±  1%        ~ (p=0.180 n=6)
RangeQuery/expr=double_exponential_smoothing(a_hundred[1d],_0.3,_0.3),steps=1000-8                                                                5.808 ±  1%                  5.820 ±  1%        ~ (p=0.180 n=6)
RangeQuery/expr=changes(a_one[1d]),steps=1000-8                                                                                                  19.33m ±  0%                 19.27m ±  1%        ~ (p=0.180 n=6)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-8                                                                                                  192.6m ±  0%                 192.4m ±  0%        ~ (p=0.240 n=6)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-8                                                                                               1.921 ±  0%                  1.927 ±  3%        ~ (p=0.093 n=6)
RangeQuery/expr=rate(a_one[1d]),steps=1000-8                                                                                                     5.853m ±  1%                 5.864m ±  1%        ~ (p=0.485 n=6)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-8                                                                                                     58.73m ±  2%                 59.01m ± 14%        ~ (p=0.310 n=6)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-8                                                                                                 584.3m ±  1%                 589.7m ±  3%   +0.93% (p=0.009 n=6)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-8                                                                                         2.661m ±  1%                 2.706m ±  2%   +1.70% (p=0.041 n=6)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-8                                                                                         26.74m ±  3%                 26.75m ±  3%        ~ (p=1.000 n=6)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-8                                                                                     269.6m ±  4%                 269.2m ±  4%        ~ (p=0.485 n=6)
RangeQuery/expr=-a_one,steps=1000-8                                                                                                              37.32µ ±  1%                 36.99µ ±  1%   -0.90% (p=0.041 n=6)
RangeQuery/expr=-a_ten,steps=1000-8                                                                                                              312.6µ ±  2%                 308.7µ ±  0%   -1.22% (p=0.002 n=6)
RangeQuery/expr=-a_hundred,steps=1000-8                                                                                                          3.096m ±  0%                 3.058m ±  0%   -1.22% (p=0.002 n=6)
RangeQuery/expr=a_one_-_b_one,steps=1000-8                                                                                                       355.3µ ± 17%                 351.0µ ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=a_ten_-_b_ten,steps=1000-8                                                                                                       3.390m ±  0%                 3.395m ±  0%        ~ (p=0.485 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-8                                                                                               37.95m ±  0%                 38.05m ±  0%        ~ (p=0.132 n=6)
RangeQuery/expr=a_one_-_b_one,steps=10000-8                                                                                                      3.509m ±  2%                 3.467m ±  0%        ~ (p=0.065 n=6)
RangeQuery/expr=a_ten_-_b_ten,steps=10000-8                                                                                                      34.58m ±  1%                 34.50m ±  1%        ~ (p=0.180 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                                                              385.0m ±  0%                 386.8m ±  2%        ~ (p=0.132 n=6)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-8                                                                                      124.5µ ±  1%                 125.0µ ±  4%        ~ (p=0.394 n=6)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                                                      1.201m ±  0%                 1.201m ±  0%        ~ (p=0.937 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                                                              14.90m ±  5%                 14.91m ±  1%        ~ (p=0.818 n=6)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-8                                                                                       186.2µ ±  0%                 185.3µ ±  0%   -0.46% (p=0.026 n=6)
RangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                                                       1.756m ±  1%                 1.746m ±  0%        ~ (p=0.065 n=6)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                                                               19.88m ±  0%                 19.96m ±  0%        ~ (p=0.065 n=6)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-8                                                                                   184.0µ ±  1%                 184.3µ ±  4%        ~ (p=0.818 n=6)
RangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-8                                                                                   1.205m ±  4%                 1.204m ±  5%        ~ (p=0.818 n=6)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                                                           14.88m ±  1%                 14.89m ±  2%        ~ (p=0.589 n=6)
RangeQuery/expr=a_one_and_b_one{l='notfound'},steps=1000-8                                                                                       117.2µ ±  1%                 116.3µ ±  1%        ~ (p=0.485 n=6)
RangeQuery/expr=a_ten_and_b_ten{l='notfound'},steps=1000-8                                                                                       448.7µ ±  0%                 451.7µ ±  1%   +0.65% (p=0.026 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=1000-8                                                                               3.817m ±  1%                 3.893m ±  1%   +1.98% (p=0.002 n=6)
RangeQuery/expr=abs(a_one),steps=1000-8                                                                                                          141.0µ ±  1%                 141.4µ ±  2%        ~ (p=0.394 n=6)
RangeQuery/expr=abs(a_ten),steps=1000-8                                                                                                          1.041m ±  2%                 1.042m ±  2%        ~ (p=0.818 n=6)
RangeQuery/expr=abs(a_hundred),steps=1000-8                                                                                                      10.97m ±  1%                 10.92m ±  0%        ~ (p=0.180 n=6)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1000-8                                                                       44.43µ ±  5%                 43.86µ ±  1%   -1.29% (p=0.002 n=6)
RangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=1000-8                                                                       316.0µ ±  2%                 314.8µ ±  0%        ~ (p=0.180 n=6)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1000-8                                                                   3.059m ±  1%                 3.045m ±  3%        ~ (p=0.394 n=6)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1000-8                                                                              41.46µ ±  1%                 41.53µ ±  0%        ~ (p=0.818 n=6)
RangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=1000-8                                                                              312.2µ ±  1%                 311.5µ ±  1%        ~ (p=0.699 n=6)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1000-8                                                                          3.036m ±  1%                 3.041m ±  2%        ~ (p=0.394 n=6)
RangeQuery/expr=sum(a_one),steps=1000-8                                                                                                          66.03µ ±  2%                 65.67µ ±  1%        ~ (p=0.190 n=6)
RangeQuery/expr=sum(a_ten),steps=1000-8                                                                                                          369.9µ ±  2%                 363.3µ ±  2%        ~ (p=0.065 n=6)
RangeQuery/expr=sum(a_hundred),steps=1000-8                                                                                                      3.453m ±  1%                 3.428m ±  1%   -0.73% (p=0.004 n=6)
RangeQuery/expr=avg(a_one),steps=1000-8                                                                                                          66.24µ ±  1%                 66.14µ ±  1%        ~ (p=0.818 n=6)
RangeQuery/expr=avg(a_ten),steps=1000-8                                                                                                          372.8µ ±  1%                 367.5µ ±  1%   -1.43% (p=0.009 n=6)
RangeQuery/expr=avg(a_hundred),steps=1000-8                                                                                                      3.502m ±  1%                 3.472m ±  4%        ~ (p=0.589 n=6)
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-8                                                                                              480.3µ ±  1%                 476.0µ ±  0%   -0.90% (p=0.002 n=6)
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-8                                                                                              3.861m ±  2%                 3.808m ±  2%   -1.38% (p=0.041 n=6)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-8                                                                                          42.52m ±  2%                 42.10m ±  1%   -1.00% (p=0.015 n=6)
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-8                                                                                             403.6µ ± 32%                 401.0µ ±  2%        ~ (p=0.240 n=6)
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-8                                                                                             3.884m ±  1%                 3.810m ±  1%   -1.90% (p=0.002 n=6)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-8                                                                                         43.63m ±  1%                 43.38m ±  1%        ~ (p=0.093 n=6)
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-8                                                                                                   403.4µ ±  2%                 399.0µ ±  1%   -1.09% (p=0.015 n=6)
RangeQuery/expr=sum_by_(l)(h_ten),steps=1000-8                                                                                                   3.894m ±  2%                 3.817m ±  1%   -1.97% (p=0.002 n=6)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-8                                                                                               43.76m ±  2%                 43.11m ±  1%   -1.48% (p=0.009 n=6)
RangeQuery/expr=sum_by_(le)(h_one),steps=1000-8                                                                                                  479.6µ ±  1%                 476.0µ ±  1%   -0.75% (p=0.041 n=6)
RangeQuery/expr=sum_by_(le)(h_ten),steps=1000-8                                                                                                  3.904m ±  1%                 3.834m ±  1%   -1.81% (p=0.002 n=6)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-8                                                                                              42.98m ±  4%                 42.20m ±  1%   -1.82% (p=0.004 n=6)
RangeQuery/expr=count_values('value',_h_one),steps=100-8                                                                                         928.3µ ±  1%                 925.4µ ±  1%        ~ (p=0.485 n=6)
RangeQuery/expr=count_values('value',_h_ten),steps=100-8                                                                                         10.79m ±  0%                 10.87m ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-8                                                                                     133.3m ±  3%                 132.3m ±  5%        ~ (p=0.699 n=6)
RangeQuery/expr=topk(1,_a_one),steps=1000-8                                                                                                      277.4µ ±  1%                 284.2µ ±  0%   +2.45% (p=0.002 n=6)
RangeQuery/expr=topk(1,_a_ten),steps=1000-8                                                                                                      671.1µ ±  2%                 672.8µ ±  2%        ~ (p=0.589 n=6)
RangeQuery/expr=topk(1,_a_hundred),steps=1000-8                                                                                                  4.222m ±  2%                 4.267m ±  3%   +1.07% (p=0.026 n=6)
RangeQuery/expr=topk(5,_a_one),steps=1000-8                                                                                                      278.0µ ±  1%                 283.7µ ±  1%   +2.06% (p=0.002 n=6)
RangeQuery/expr=topk(5,_a_ten),steps=1000-8                                                                                                      1.264m ±  1%                 1.268m ±  1%        ~ (p=0.132 n=6)
RangeQuery/expr=topk(5,_a_hundred),steps=1000-8                                                                                                  7.271m ±  1%                 7.321m ±  1%        ~ (p=0.180 n=6)
RangeQuery/expr=limitk(1,_a_one),steps=1000-8                                                                                                    268.4µ ±  1%                 273.8µ ±  1%   +1.98% (p=0.002 n=6)
RangeQuery/expr=limitk(1,_a_ten),steps=1000-8                                                                                                    572.4µ ±  4%                 581.4µ ±  0%        ~ (p=0.065 n=6)
RangeQuery/expr=limitk(1,_a_hundred),steps=1000-8                                                                                                3.345m ±  2%                 3.488m ±  1%   +4.28% (p=0.002 n=6)
RangeQuery/expr=limitk(5,_a_one),steps=1000-8                                                                                                    268.5µ ±  3%                 274.4µ ±  1%   +2.21% (p=0.041 n=6)
RangeQuery/expr=limitk(5,_a_ten),steps=1000-8                                                                                                    967.6µ ±  3%                 985.9µ ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=limitk(5,_a_hundred),steps=1000-8                                                                                                3.764m ±  2%                 3.877m ±  1%   +3.01% (p=0.002 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_one),steps=1000-8                                                                                             226.2µ ±  1%                 225.0µ ±  2%        ~ (p=0.065 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_ten),steps=1000-8                                                                                             1.233m ±  1%                 1.242m ±  1%        ~ (p=0.093 n=6)
RangeQuery/expr=limit_ratio(0.1,_a_hundred),steps=1000-8                                                                                         9.599m ±  1%                 9.691m ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_one),steps=1000-8                                                                                             227.8µ ±  2%                 224.5µ ±  1%   -1.49% (p=0.015 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_ten),steps=1000-8                                                                                             1.607m ±  8%                 1.649m ±  1%        ~ (p=0.065 n=6)
RangeQuery/expr=limit_ratio(0.5,_a_hundred),steps=1000-8                                                                                         14.19m ±  1%                 14.32m ±  0%   +0.95% (p=0.002 n=6)
…